removed commented out css + moved parseNoteDate to utility.js
