+++ source/styling/index.css
@@ -330,13 +330,7 @@ main {
 
 .note-more {
   height: 3rem;
-  /* width: auto; */
-  /* border-radius: 100%; */
-  /* filter: invert(70%); */
-  /* z-index: 10; */
   transition: transform 0.3s;
-
-  /* background: red; */
   width: 3rem;
   border-bottom-right-radius: 0.4rem;
   border-top-left-radius: 0.4rem;
